Update products listed on enter key
