+++ components/searchBar.tsx
@@ -2,10 +2,10 @@
 
 import React from 'react';
 
-export default function SearchBar({query, handleInputChange}) {
+export default function SearchBar({query, handleSearchBarInputChange, handleFormSubmit}) {
 
     return(
-        <form className="mx-3 max-w-md h-10">
+        <form className="mx-3 max-w-md h-10" onSubmit={handleFormSubmit}>
         <label
             htmlFor="default-search"
             className="mb-2 text-sm font-medium text-gray-900 sr-only"
@@ -36,14 +36,8 @@ export default function SearchBar({query, handleInputChange}) {
             className="block w-96 p-2 ps-10 text-sm text-gray-900 border border-gray-300 rounded-full bg-white focus:ring-gray-500 focus:border-gray-500"
             placeholder="Search"
             value={query}
-            onChange={handleInputChange}
+            onChange={handleSearchBarInputChange}
             />
-            {/* <button
-            type="submit"
-            className="text-white absolute end-2.5 bottom-2.5 bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-4 py-2 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
-            >
-            Search
-            </button> */}
         </div>
         </form>
     )
