+++ package.json
@@ -18,7 +18,8 @@
     "next": "14.1.0",
     "react": "^18",
     "react-bootstrap": "^2.10.1",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "react-toastify": "^10.0.5"
   },
   "devDependencies": {
     "@types/node": "20.11.29",
